Move OracleParamsTLV to its own file
